requirements.txt
pandas==2.1.4
python-dotenv==1.0.0 